added dockerile and dockercompose - currently using serverOld.js and not the new express stuff
