+++ .gitignore
@@ -7,6 +7,9 @@
 # Only exists if Bazel was run
 /bazel-out
 
+# temp_passwords
+.env
+
 # dependencies
 /node_modules
