+++ server.js
@@ -1,41 +1,11 @@
-// Get dependencies
-const express = require('express');
-const path = require('path');
 const http = require('http');
-const bodyParser = require('body-parser');
+const app = require('./server/express/app');
 
-// Get our API routes
-const api = require('./server/routes/api');
-
-const app = express();
-
-// Parsers for POST data
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({extended: false}));
-
-// Point static path to dist (folder where build files are located) might have to change this name for your computer
-app.use(express.static(path.join(__dirname, 'dist/demosite')));
-
-// Set our api routes
-app.use('/api', api);
-
-// Catch all other routes and return the index file
-app.get('*', (req, res) => {
-  res.sendFile(path.join(__dirname, 'dist/demosite/index.html'));
-});
-
-/**
- * Get port from environment and store in Express.
- */
+/* set up the port variable*/
 const port = process.env.PORT || '3000';
-app.set('port', port);
+app.set("port",port);
 
-/**
- * Create HTTP server.
- */
+
+/*create server using express app and listen on port*/
 const server = http.createServer(app);
-
-/**
- * Listen on provided port, on all network interfaces.
- */
-server.listen(port, () => console.log(`API running on localhost:${port}`));
+server.listen(port);
