+++ server/express/app.js
@@ -8,6 +8,7 @@ const path = require('path');
 const mongoDBConnect = "mongodb+srv://Hari:G8F1P3MeLb77pV48@cluster0.9hjpe.mongodb.net/myFirstDatabase?retryWrites=true&w=majority";
 
 /*init an express middleware*/
+const api = require('./routes/api');
 const app = express();
 
 /*use cors to allow cross origin resource*/
@@ -20,8 +21,9 @@ app.use(bodyParser.urlencoded({extended: false}))
 
 app.use(express.static(path.join(__dirname, '../../dist/demosite')));
 // Catch all other routes and return the index file
+app.use('/api', api);
 app.get('*', (req, res) => {
-  res.sendFile(path.join(__dirname, '../../dist/demosite/index.html'));
+  res.sendFile(path.join(__dirname, '../../dist/index.html'));
 });
 
 /*connect to mongoDb*/
