+++ server/express/routes/api.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+/* GET api listing. */
+router.get('/', (req, res) => {
+  res.send('api works');
+});
+
+module.exports = router;
